fix: added avatar to horizontal cards (related-articles, features-articles, content-list) chore: structural refactoring, extraction of article.js
