+++ aemedge/blocks/featured-articles/featured-articles.js
@@ -3,15 +3,15 @@ import ffetch from '../../scripts/ffetch.js';
 import { ul } from '../../scripts/dom-builder.js';
 import PictureCard from '../../libs/pictureCard/pictureCard.js';
 import { formatDate } from '../../scripts/utils.js';
-import { asEntry } from '../../scripts/profile.js';
+import { allAuthorEntries, authorEntry } from '../../scripts/article.js';
 
-function getPictureCard(article, placeholders) {
+function getPictureCard(article, placeholders, authEntry) {
   const {
-    author, 'content-type': type, image, path, title, priority, description,
+    'content-type': type, image, path, title, priority, description,
   } = article;
   const tagLabel = placeholders[toCamelCase(priority)] || '';
   const info = `Updated on ${formatDate(article.publicationDate * 1000)}`;
-  return new PictureCard(title, path, type, info, asEntry(author), image, tagLabel, description);
+  return new PictureCard(title, path, type, info, authEntry, image, tagLabel, description);
 }
 
 export default async function decorateBlock(block) {
@@ -23,9 +23,10 @@ export default async function decorateBlock(block) {
       .filter(filter)
       .all();
     const placeholders = await fetchPlaceholders();
+    const authEntries = await allAuthorEntries(articleStream);
     const cardList = ul();
     articleStream.forEach((article) => {
-      const card = getPictureCard(article, placeholders);
+      const card = getPictureCard(article, placeholders, authorEntry(article, authEntries));
       cardList.append(card.render(horizontal));
     });
     block.append(cardList);
